add adoption component
