+++ client/src/components/meeting.js
@@ -33,29 +33,23 @@ import "../css/meeting.css";
  */
 
 function Meetings(props) {
-  if (props.status) {
-    return (
-      <div className="meeting-card">
-        <div className="title">
-          <h2>{props.title}</h2>
-        </div>
-        <div className="break" />
-        <div className="text-card">{props.textCard}</div>
-        <div className="image-card">
-          <div className="status">{props.status}</div>
-        </div>
-      </div>
-    );
-  }
   return (
     <div className="meeting-card">
       <div className="title">
         <h2>{props.title}</h2>
       </div>
-      <div className="break" />
-      <div className="text-card">{props.textCard}</div>
-      <div className="image-card">
-        <img className="image" src={props.imagePath} alt="" />
+      <div className="meeting-card-content">
+        <div className="text-card">{props.textCard}</div>
+        {props.status && (
+          <div className="image-card">
+            <div className="status">{props.status}</div>
+          </div>
+        )}
+        {props.imagePath && (
+          <div className="image-card">
+            <img className="image" src={props.imagePath} alt="" />
+          </div>
+        )}
       </div>
     </div>
   );
