+++ client/src/contexts/ApplicationContext.js
@@ -1,0 +1,5 @@
+import React from "react";
+
+const ApplicationContext = React.createContext({});
+
+export default ApplicationContext;
