+++ client/src/css/defaultbody.css
@@ -15,6 +15,9 @@
   flex-direction: column;
   padding: 50px 40px;
   overflow: auto;
+  width: 100%;
+  height: 100%;
+  box-sizing: border-box;
 }
 
 @media only screen and (max-width: 750px) {
