+++ client/src/css/meeting.css
@@ -1,30 +1,30 @@
 .meeting-card {
+  display: flex;
   border-radius: 15px;
   background: #000000;
-  position: relative;
   display: flex;
   text-align: center;
-  flex-direction: row;
-  justify-content: left;
-  width: 80%;
-  height: 40%;
-  max-width: max-content;
-  max-height: max-content;
+  flex-direction: column;
   overflow: auto;
-  margin: 5px;
-  flex-wrap: wrap;
   text-align: center;
   padding-left: 2%;
   padding-right: 2%;
   padding-top: 5px;
   padding-bottom: 2%;
-  overflow-x: hidden;
-  object-fit: contain;
-  word-wrap: break-word;
+  box-sizing: border-box;
+  height: calc(100% - 40px);
+  width: 100%;
 }
 
 h2 {
   font-size: 2em;
+}
+
+.meeting-card-content {
+  display: flex;
+  flex: 1;
+  flex-direction: row;
+  justify-content: center;
 }
 
 .title {
@@ -34,24 +34,18 @@ h2 {
 }
 
 .text-card {
-  margin-top: 0px;
-  /* margin-right: 10px; */
   background: #fff;
   flex: 2;
   border-radius: 15px;
-  width: 100%;
   padding: 2%;
   text-align: left;
-  /* padding-left: 5%; */
+  max-width: 50%;
 }
 
 .image-card {
-  margin-top: 0px;
   margin-left: 2vw;
   flex: 2;
   border-radius: 15px;
-  /* width:auto;
-    height:auto; */
   width: 100%;
   height: 100%;
 }
@@ -76,13 +70,15 @@ h2 {
   .meeting-card {
     display: flex;
     flex-direction: column;
-    /* width: auto;
-    height: auto; */
     padding-left: 4%;
     padding-right: 5%;
     padding-top: 5px;
     padding-bottom: 3%;
     font-size: 1.2em;
+  }
+
+  .meeting-card-content {
+    flex-direction: column;
   }
 
   h2 {
@@ -92,15 +88,12 @@ h2 {
   .text-card {
     min-height: 50vw;
     width: auto;
-    /* height: auto; */
-    /* margin-bottom: 30px; */
     margin: 1% 1% 4% 1%;
+    max-width: unset;
   }
 
   .image-card {
     flex: 1;
-    /* width: auto;
-    height: auto; */
     margin: 1%;
   }
